cosfunc {cvv}	R Documentation

Create a function to compute lagged cosines of velocity vectors

Description

Given a dyadic motion object produced by make.fd, along with a fixed lag, outputs a function mapping time points to lagged cosine of velocity vectors.

Usage

cosfunc(obj = NULL, x0, y0, z0, x1, y1, z1, lag = 0)

Arguments

obj
A dyadic motion object outputted by make.fd.

x0, y0, z0, x1, y1, z1
Functional data objects (see fd) representing the three coordinates for persons 0 and 1. These are extracted from obj if the latter is provided.

lag
Lag for person 1 with respect to person 0. E.g., if lag=0.1, person 1 is assumed to lag behind person 0 by 0.1 second.

Value

A function that inputs a vector of times, and outputs the cosines between the two 3D velocity vectors at each of these times, at the given lag.

See Also

multilag, which depends on this function. Most users would probably use multilag rather than calling cosfunc directly.

[Package cvv version 0.1 Index]