make.fd {cvv}	R Documentation

Processing of dyadic 3D motion data

Description

Inputs times and corresponding positions in 3D space for two individuals, and uses the fda package to create functional data objects representing both persons' data with respect to a B-spline basis.

Usage

make.fd(tvec, p0, p1, nbasis = 300, lambda = NULL)

Arguments

tvec
Vector of times

p0, p1
Matrices of 3D positions for persons 0 and 1 at the times given by tvec; each should have dimension length(tvec) by 3

nbasis
Dimension of B-spline basis used to represent each person's data along each dimension

lambda
Tuning parameter for penalized smoothing with respect to the given B-spline basis.

Value

A list with elements

x0,y0,z0
Functional data objects (see fd) representing the three coordinates for person 0

x1,y1,z1
Same, for person 1

bsb
B-spline basis object (see create.bspline.basis)

Examples

# see ?syncsham

[Package cvv version 0.1 Index]