scat3d {cvv}	R Documentation

3D plotting of motion data

Description

Inputs a dyadic motion data produced by make.fd, and creates a 3D plot of the data for one member of the dyad.

Usage

scat3d(obj, person, npoints = 1001, deriv = 0, tmin = NULL,
  tmax = NULL, cex = 0.5, col = rainbow(npoints, end = 5/6),
  plotter = "rgl", common.width = FALSE, ...)

Arguments

obj
A dyadic motion object outputted by make.fd

person
Which person's data should be plotted: 0 or 1

npoints
Number of (equally spaced) points to include in the plot

deriv
Derivative to plot (e.g., 1 for velocity, 2 for acceleration); the default, 0, plots the positions

tmin, tmax
Lower and upper ends of the time range

cex
Magnification factor for the points

col
Color; the default gives a rainbow plot

plotter
Character vector describing the plotting function to use: either "plot3d" or "scatterplot3d"

common.width
Logical: Should the plotting interval along each of the 3 dimensions be of equal width?

...
Arguments passed to the plotting function.

Value

A list with components

tseq
Sequence of times

xfd,yfd,zfd
Functional data objects (see fd) for the x, y, and z-coordinates

Examples

# 3D plots as in Fig. 6 of Reiss, Gvirts et al.
data(sync.fd)
rnge <- c(61,62.2)  
par(mfrow=1:2)
pers0 <- scat3d(sync.fd, person=0, tmin=rnge[1], tmax=rnge[2], 
           plotter="scatterplot3d", main="Leader", xlab='x', ylab='y', zlab='z')
pers1 <- scat3d(sync.fd, person=1, tmin=rnge[1], tmax=rnge[2], 
           plotter="scatterplot3d", main="Follower", xlab='x', ylab='y', zlab='z')

[Package cvv version 0.1 Index]