syncsham {cvv}	R Documentation

Synchronized 3D motion for a dyad

Description

Times and positions for two individuals moving synchronously during sham brain stimulation.

Format

This data set consists of three objects:

tvec.

A vector of 41510 time points

p0.,p1.

Two 41510 x 3 matrices, giving 3D positions for two individuals at the times given by tvec.

See Also

make.fd converts these data to sync.fd, which is in the required format for computing CVV via function cosfunc

Examples

## Not run: 
data(syncsham)
tmp <- make.fd(tvec., p0., p1.)   # may take a while...
data(sync.fd)
all.equal(tmp, sync.fd)

## End(Not run)

[Package cvv version 0.1 Index]